Fix: Created utils folder, moved memory.py
