+++ bot/utils/memory.py
