+++ bot/utils/notion.py
@@ -1,0 +1,2 @@
+from notion_client import AsyncClient
+
